style: fixed linting errors; added SwiftLint as Xcode run script
