+++ Sources/OnLaunch/API/ApiMessageDto.swift
@@ -1,0 +1,7 @@
+struct ApiMessageDto: Decodable {
+    let id: Int
+    let blocking: Bool
+    let title: String
+    let body: String
+    let actions: [ApiActionDto]
+}
