+++ Sources/OnLaunch/API/ApiMessagesResponseDto.swift
@@ -1,0 +1,1 @@
+typealias ApiMessagesResponseDto = [ApiMessageDto]
